requirements.txt
# langchain
# llama-cpp-python
huggingface_hub[cli]
colorama
uvicorn
fastapi
sse_starlette
safetensors
# exllama @ git+https://github.com/jllllll/exllama
sentencepiece>=0.1.99
# flash-attn
pypdf
pycryptodome
# flash-attn --no-build-isolation
# exllamav2
sqlmodel
py7zr
eciespy
python-multipart
ray[serve]
transformers
markdownify
pypdf2
openai
tiktoken
ocrmypdf
google-search-results
sentence-transformers
lm-format-enforcer
vllm
readability-lxml
selenium
psycopg2-binary
fastapi-login
FlagEmbedding
pgvector
python-jose
passlib
lxml_html_clean
exllamav2
jsonpatch
jsonpath-ng
peft
# pluggy
# rich

pypdfium2

psycopg2-binary==2.9.9
sqlalchemy==2.0.35
sqlitedict==2.1.0
sqlmodel==0.0.22


marker-pdf==0.3.9
surya-ocr==0.6.11
